requirements.txt
# Framework Web
#Flask==3.1.0
Flask

# Data Processing
#pandas==2.2.3
pandas>=2.3,<3
#numpy==2.1.3
numpy>=2.0,<3
openpyxl==3.1.5

# Survival Analysis
#lifelines==0.29.0
lifelines

# Scientific Computing e Ottimizzazione
scipy>=1.15.0,<2

# Plotting
matplotlib==3.9.2

# Server WSGI (opzionale, per produzione)
gunicorn==23.0.0

# Autenticazione e Forms
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Security
Werkzeug==3.0.1

# postgres
psycopg2-binary>=2.9,<3
Flask-SQLAlchemy>=3.1
python-dotenv>=1.0.0

